consolidating controllers into 1 file
