+++ uap_tracker/controller.py
@@ -18,9 +18,18 @@ from uap_tracker.frame_processor import FrameProcessor
 from uap_tracker.dense_optical_flow import DenseOpticalFlow
 from uap_tracker.background_subtractor_factory import BackgroundSubtractorFactory
 
-class CameraStreamController():
+class Controller():
+
+    def __init__(self):
+        pass
+
+    def run(self):
+        pass
+
+class CameraStreamController(Controller):
 
     def __init__(self, camera, video_tracker, minute_interval=10):
+        super().__init__()
 
         self.camera = camera
         self.video_tracker = video_tracker
@@ -82,3 +91,47 @@ class CameraStreamController():
                     self.video_tracker.finalise()
                     self.running = False
                     break
+
+class VideoPlaybackController(Controller):
+
+    def __init__(self, capture, video_tracker):
+        super().__init__()
+
+        self.capture = capture
+        self.video_tracker = video_tracker
+
+    def run(self):
+
+        success, init_frame = self.capture.read()
+        if not success:
+            print(f"Could not open video stream")
+            sys.exit()
+
+        frame_count = 0
+        fps = 0
+        background_subtractor = BackgroundSubtractorFactory.Select(enable_cuda=self.video_tracker.settings['enable_cuda'], sensitivity=self.video_tracker.settings['detection_sensitivity'])
+
+        dense_optical_flow = None
+        if self.video_tracker.settings['calculate_optical_flow']:
+            dense_optical_flow = DenseOpticalFlow.Select(enable_cuda=self.video_tracker.settings['enable_cuda'], width=480, height=480)
+
+        with FrameProcessor.Select(
+            settings=self.video_tracker.settings,
+            dense_optical_flow=dense_optical_flow,
+            background_subtractor=background_subtractor) as processor:
+
+            # Mike: Initialise the tracker and processor
+            self.video_tracker.initialise(processor, init_frame)
+
+            while cv2.waitKey(1) != 27:  # Escape
+                success, frame = self.capture.read()
+                if success:
+                    timer = cv2.getTickCount()
+                    self.video_tracker.process_frame(processor, frame, frame_count, fps)
+                    # Calculate Frames per second (FPS)
+                    fps = cv2.getTickFrequency() / (cv2.getTickCount() - timer)
+                    frame_count += 1
+                else:
+                    break
+
+            self.video_tracker.finalise()
